+++ Tests/FluentAssertions.Equivalency.Specs/MemberMatchingSpecs.cs
@@ -440,7 +440,7 @@ namespace FluentAssertions.Equivalency.Specs
 
             var expectation = new
             {
-                Property2 = 2, 
+                Property2 = 2,
                 Ignore = 3
             };
 
@@ -463,7 +463,7 @@ namespace FluentAssertions.Equivalency.Specs
 
             var expectation = new
             {
-                Property2 = 2, 
+                Property2 = 2,
                 Ignore = 3
             };
 
@@ -473,6 +473,44 @@ namespace FluentAssertions.Equivalency.Specs
                     .ExcludingMissingMembers()
                     .WithMapping("Property2", "Property1")
                 );
+        }
+
+        [Fact]
+        public void Can_map_members_of_a_root_collection()
+        {
+            // Arrange
+            var entity = new Entity
+            {
+                EntityId = 1,
+                Name = "Test"
+            };
+
+            var dto = new EntityDto
+            {
+                Id = 1,
+                Name = "Test"
+            };
+
+            var entityCol = new[] { entity };
+            var dtoCol = new[] { dto };
+
+            // Act / Assert
+            dtoCol.Should().BeEquivalentTo(entityCol, c =>
+                c.WithMapping<EntityDto>(s => s.EntityId, d => d.Id));
+        }
+
+        private class Entity
+        {
+            public int EntityId { get; init; }
+
+            public string Name { get; init; }
+        }
+
+        private class EntityDto
+        {
+            public int Id { get; init; }
+
+            public string Name { get; init; }
         }
 
         internal class ParentOfExpectationWithProperty2